requirements.txt
aiogram==2.23.1
async-timeout==4.0.2
charset-normalizer==2.1.1

idna==3.4
magic-filter==1.0.9



python-decouple==3.6
pytz==2022.6
requests==2.28.1

typing_extensions==4.4.0
urllib3==1.26.13

